scm.manager fix all detached objects/process - in case anything unexpected happened close/reopen a new session - due that `scoped_session` always give a thread independent session

+++ anm/careas/scm/manager.py
@@ -40,9 +40,6 @@ class ProcessManagerClass(dict):
     * value : `scm.Processo` object
     """    
     def __init__(self, debug=False): 
-        """
-        * save_on_set: save on database on set 
-        """        
         super().__init__()      
         self._engine = create_engine(f"sqlite:///{config['scm']['process_storage_file']+'.db'}")                    
         self.__session = scoped_session(sessionmaker(bind=self._engine))
@@ -70,12 +67,18 @@ class ProcessManagerClass(dict):
         if I had to open/close the session every time, instead of letting 
         `scoped_session` close it at the end of the thread.
         """
-        return self.__session()
+        with self.lock:
+            return self.__session()
 
-    def __delitem__(self, key : str):        
-        self.session.delete(self[key].db)
-        self.session.commit()
-        super().__delitem__(key)
+    def __delitem__(self, key : str):    
+        """
+        this is called when del ProcessManager[key] 
+        be careful with this 
+        """
+        if self[key]:    
+            self.session.delete(self[key].db)
+            self.session.commit()
+            super().__delitem__(key)
     
     def __getitem__(self, key : str) -> Processo:
         """
@@ -87,6 +90,14 @@ class ProcessManagerClass(dict):
             key = fmtPname(key)    
             if key in self:
                 processo = super().__getitem__(key)   
+                # if a process was previouly added to the manager but 
+                # its thread was closed or anything happened it got
+                # detached. We need to add it to this/new/current session
+                session_ = object_session(processo.db) # original session
+                if session_ is not self.session: # is not 'this' session   
+                    if session_ is not None: # close it if wasn't already
+                        session_.close()            
+                    self.session.add(processo.db)  # add to current/new one
                 return processo
             else:
                 processodb = self.session.query(Processodb).filter_by(name=key).first()
@@ -111,48 +122,35 @@ class ProcessManagerClass(dict):
             pdb = self[name]
             if pdb is None:
                 raise KeyError(name)
-            session_ = object_session(pdb.db)
-            if session_ is not self.session:    
-                if session_ is not None: 
-                    session_.close()            
-                self.session.add(pdb.db)     
-            pdb.db.dados[key] = value
+            pdb.db.dados.update({key : value}) 
             self.session.commit()
             self.session.close()   
 
 
     def getDados(self, name):
         """
-        For work in detached mode - Flask WSGI Request usage (more bellow)
+        For work in detached mode - always close the session after
+        E.g. Flask WSGI Request usage (more bellow)    
         return Processo.dados
         """
         with self.lock:            
             pdb = self[name]
             if pdb is None:
                 raise KeyError(name)
-            session_ = object_session(pdb.db)
-            if session_ is not self.session:    
-                if session_ is not None: 
-                    session_.close()            
-                self.session.add(pdb.db)  
             dados = pdb.dados
             self.session.close()
             return dados
 
     def getAttr(self, name, attr):
         """
-        For work in detached mode - Flask WSGI Request usage (more bellow)
+        For work in detached mode - always close the session after
+        E.g. Flask WSGI Request usage (more bellow)    
         return Processo.attr
         """
         with self.lock:            
             pdb = self[name]
             if pdb is None:
                 raise KeyError(name)
-            session_ = object_session(pdb.db)
-            if session_ is not self.session:    
-                if session_ is not None: 
-                    session_.close()            
-                self.session.add(pdb.db)  
             data = getattr(pdb, attr, None)
             self.session.close()
             return data
@@ -162,12 +160,12 @@ class ProcessManagerClass(dict):
         Get `.all()` processos querying with sqlalchemy filter 
         example:
         getting processes with 'clayers' key in 'dados->iestudo'
-        ProcessManager.getwithFilter( text("dados->'iestudo' ? 'clayers'"))
+        ProcessManager.getwithFilter( text("dados->'estudo ? 'clayers'"))
         hence can't close session otherwise processes will be detached 
         raising DetachedInstanceError
         """
-        with self.lock:                
-            processes = session.query(Processodb).filter(filter_condition).all()            
+        with self.lock:             
+            processes = self.session.query(Processodb).filter(filter_condition).all()            
             return processes
